cs: refine vm golden tests
