+++ compiler/x/cs/vm_golden_test.go
@@ -4,6 +4,7 @@ package cscode_test
 
 import (
 	"bytes"
+	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -15,6 +16,33 @@ import (
 	"mochi/runtime/vm"
 	"mochi/types"
 )
+
+func stripHeaderVM(b []byte) []byte {
+	if i := bytes.IndexByte(b, '\n'); i != -1 && bytes.HasPrefix(b, []byte("// Generated by Mochi")) {
+		return b[i+1:]
+	}
+	return b
+}
+
+func repoRoot(t *testing.T) string {
+	t.Helper()
+	dir, err := os.Getwd()
+	if err != nil {
+		t.Fatal("cannot determine working directory")
+	}
+	for i := 0; i < 10; i++ {
+		if _, err := os.Stat(filepath.Join(dir, "go.mod")); err == nil {
+			return dir
+		}
+		parent := filepath.Dir(dir)
+		if parent == dir {
+			break
+		}
+		dir = parent
+	}
+	t.Fatal("go.mod not found")
+	return ""
+}
 
 func TestCSCompiler_VMValid_Golden(t *testing.T) {
 	if err := cscode.EnsureDotnet(); err != nil {
@@ -45,9 +73,11 @@ func TestCSCompiler_VMValid_Golden(t *testing.T) {
 			}
 			wantCodePath := filepath.Join(outDir, name+".cs")
 			wantOutPath := filepath.Join(outDir, name+".out")
+			gotCode := bytes.TrimSpace(stripHeaderVM(code))
 			if want, err := os.ReadFile(wantCodePath); err == nil {
-				if got := bytes.TrimSpace(code); !bytes.Equal(got, bytes.TrimSpace(want)) {
-					t.Errorf("generated code mismatch for %s.cs\n\n--- Got ---\n%s\n\n--- Want ---\n%s", name, got, bytes.TrimSpace(want))
+				wantCode := bytes.TrimSpace(stripHeaderVM(want))
+				if !bytes.Equal(gotCode, wantCode) {
+					t.Errorf("generated code mismatch for %s.cs\n\n--- Got ---\n%s\n\n--- Want ---\n%s", name, gotCode, wantCode)
 				}
 			}
 			dir := t.TempDir()
@@ -80,8 +110,41 @@ func TestCSCompiler_VMValid_Golden(t *testing.T) {
 					}
 				}
 			}
-			_ = os.WriteFile(wantCodePath, code, 0644)
+			_ = os.WriteFile(wantCodePath, gotCode, 0644)
 			_ = os.WriteFile(wantOutPath, gotOut, 0644)
 		})
 	}
+}
+
+func TestMain(m *testing.M) {
+	code := m.Run()
+	updateReadme()
+	os.Exit(code)
+}
+
+func updateReadme() {
+	root := repoRoot(&testing.T{})
+	srcDir := filepath.Join(root, "tests", "vm", "valid")
+	outDir := filepath.Join(root, "tests", "machine", "x", "cs")
+	files, _ := filepath.Glob(filepath.Join(srcDir, "*.mochi"))
+	total := len(files)
+	compiled := 0
+	var lines []string
+	for _, f := range files {
+		name := strings.TrimSuffix(filepath.Base(f), ".mochi")
+		mark := "[ ]"
+		if _, err := os.Stat(filepath.Join(outDir, name+".out")); err == nil {
+			compiled++
+			mark = "[x]"
+		}
+		lines = append(lines, fmt.Sprintf("- %s %s", mark, name))
+	}
+	var buf bytes.Buffer
+	buf.WriteString("# C# Machine Output\n\n")
+	buf.WriteString("This directory holds C# source generated from the Mochi programs in `tests/vm/valid`. Each compiled program has a `.cs` file and the expected output in a matching `.out`. If the compiler failed a `.error` file will be present instead.\n\n")
+	fmt.Fprintf(&buf, "Compiled programs: %d/%d\n\n", compiled, total)
+	buf.WriteString("Checklist:\n\n")
+	buf.WriteString(strings.Join(lines, "\n"))
+	buf.WriteString("\n")
+	_ = os.WriteFile(filepath.Join(outDir, "README.md"), buf.Bytes(), 0o644)
 }
