+++ tests/machine/x/cs/print_hello.cs
@@ -1,7 +1,9 @@
 using System;
 
-class Program {
-    static void Main() {
+class Program
+{
+    static void Main()
+    {
         Console.WriteLine("hello");
     }
 }
